fix: немедленное обновление UI после перемещения файлов и AI операций
Проблемы:
- После перемещения файла через MoveDialog он исчезал с экрана до обновления
- После выполнения AI команд изменения не отражались сразу в UI
- invalidateQueries помечал кэш устаревшим, но не обновлял данные

Решение:
- Добавлен refetchQueries после invalidateQueries в useMoveNode
- Передача oldParentId для обновления обеих родительских папок
- Обновление дерева и списка детей немедленно после мутации
- Аналогичное исправление для AI команд в AIModal
- Исправлен SQL запрос: добавлен ::int каст для substring в postgres_node_repo

+++ storage-service/internal/infrastructure/database/postgres_node_repo.go
@@ -264,7 +264,7 @@ WHERE id = $1`
 	// depth = depth + delta
 	delta := n.Depth() - oldPath.Depth()
 	const updDescendants = `UPDATE nodes
-SET path = $1 || substring(path FROM $4),
+SET path = $1 || substring(path FROM $4::int),
     depth = depth + $5,
     updated_at = $6
 WHERE path LIKE $2 AND id <> $3`
